requirements.txt
pillow
torch
torchvision
tqdm
numpy==1.26.4
opencv-python